fix: :rocket: Update merchant bank info to merchant bank accounts.
