+++ packages/affiliation-merchant-bank-accounts/README.md
@@ -1,4 +1,4 @@
-# Sling Merchant Bank Info Table Component.
+# Sling Merchant Bank accounts Table Component.
 
 This component renders a table element with the merchant bank account information within. It uses table basic component as content wrapper.
 
@@ -8,10 +8,10 @@ This is how you set the component into the project html structure
 
 ```html
     <!-- component wrapper page location -->
-    <affiliation-merchant-bank-info
+    <affiliation-merchant-bank-accounts
         editable
         affiliationCode="">
-    </affiliation-merchant-bank-info>
+    </affiliation-merchant-bank-accounts>
 ```
 
 The *affiliationCode* attribute is mandatory for the component to make the correct request to the integrated API and populate it with data.
