+++ packages/affiliation-merchant-bank-accounts/public/index.html
@@ -6,13 +6,13 @@
     <meta http-equiv="x-ua-compatible" content="ie=edge">
     <script src="../../../scripts/helpers/injectDependencies.js"></script>
     <link rel="stylesheet" href="index.css">
-    <title>sling-merchant-bank-info</title>
+    <title>sling-merchant-bank-accounts</title>
   </head>
   <body>
     <!-- the component needs to receive a valid stonecode and auth token to correct load data -->
-    <affiliation-merchant-bank-info
+    <affiliation-merchant-bank-accounts
       editable
       affiliationCode="">
-    </affiliation-merchant-bank-info>
+    </affiliation-merchant-bank-accounts>
   </body>
 </html>
