+++ packages/affiliation-merchant-bank-accounts/src/index.css
@@ -1,0 +1,2 @@
+@import url('./assets/MerchantBankAccounts.css');
+@import url('./assets/host.css');
