+++ packages/affiliation-merchant-bank-accounts/src/index.js
@@ -1,0 +1,6 @@
+import { registerComponent } from 'sling-helpers';
+import { SlingElement } from 'sling-framework';
+import { AffiliationMerchantBankAccounts } from './component/MerchantBankAccounts.js';
+
+registerComponent('affiliation-merchant-bank-accounts',
+  AffiliationMerchantBankAccounts(SlingElement));
