--- packages/affiliation-merchant-bank-info/src/index.css
@@ -1,2 +1,0 @@
-@import url('./assets/MerchantBankInfo.css');
-@import url('./assets/host.css');
